@@ -27,15 +27,15 @@
     ]
   },
   "dependencies": {
-    "@chakra-ui/icons": "^1.1.7",
-    "@chakra-ui/react": "^1.8.8",
+    "@chakra-ui/icons": "^2.0.17",
+    "@chakra-ui/react": "^2.4.9",
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.9.3",
     "framer-motion": "^3.10.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "linguabook-components": "^1.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "readline": "^1.3.0",
-    "web-vitals": "^3.1.1",
-    "linguabook-components": "^1.0.0"
+    "web-vitals": "^3.1.1"
   }
 }
