@@ -38,8 +38,8 @@
     ]
   },
   "dependencies": {
-    "@chakra-ui/icons": "^1.1.7",
-    "@chakra-ui/react": "^1.8.8",
+    "@chakra-ui/icons": "^2.0.17",
+    "@chakra-ui/react": "^2.4.9",
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.9.3",
     "axios": "^1.2.2",
@@ -57,8 +57,7 @@
     "react-icons": "^4.7.1",
     "react-infinite-scroll-component": "^6.1.0",
     "react-responsive": "^9.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-router-ga": "^1.2.3",
+    "react-router-dom": "^6.8.0",
     "readline": "^1.3.0",
     "recoil": "^0.7.6",
     "swiper": "^8.4.5",
@@ -77,6 +76,6 @@
     "@types/react-router-dom": "^5.3.3"
   },
   "peerDependencies": {
-    "react": "^17.0.2"
+    "react": "^18.2.0"
   }
 }
